legere modification du comportement de l'api (ajout de verifs)

+++ ApiLaravel/app/Http/Controllers/GoodieController.php
@@ -24,20 +24,38 @@ class GoodieController extends Controller
         return Goodie::create($request->all());
     }
 
-    public function show(Goodie $goodie)
+    public function show($id)
     {
-        return $goodie;
+        $goodie = Goodie::find($id);
+
+        if (!$goodie) {
+            return response()->json(['message' => 'Goodie non trouvé'], 404);
+        }
+
+        return response()->json($goodie, 200);
     }
 
-    public function update(Request $request, Goodie $goodie)
+    public function update(Request $request, $id)
     {
+        $goodie = Goodie::find($id);
+
+        if (!$goodie) {
+            return response()->json(['message' => 'Goodie non trouvé'], 404);
+        }
+
         $goodie->update($request->all());
-        return $goodie;
+        return response()->json($goodie, 200);
     }
 
-    public function destroy(Goodie $goodie)
+    public function destroy($id)
     {
+        $goodie = Goodie::find($id);
+
+        if (!$goodie) {
+            return response()->json(['message' => 'Goodie non trouvé'], 404);
+        }
+
         $goodie->delete();
-        return response()->json(['message' => 'Goodie supprimé']);
+        return response()->json(['message' => 'Goodie supprimé'], 200);
     }
 }
